Fix base_map because I accidentally made it too inconvenient
