+++ src/data/include/data/functional/map.hpp
@@ -81,79 +81,7 @@ namespace data {
         std::ostream inline &write (std::ostream &o, const M &m) {
             return o << m.values ();
         }
-        
-        // we find that it is useful to have a map that also has other
-        // attributes. This type is useful for creating a map that can
-        // be extended.
 
-        template <typename derived, typename K, typename V, functional::map<K, V> M>
-        struct base_map : M {
-            using M::map;
-
-            // the derived type needs to inheret these constructors.
-            base_map (M &&rb);
-            base_map (const M &rb);
-
-            derived insert (const K &k, const V &v) const;
-            derived insert (const entry<K, V> &e) const;
-            derived insert (const K &k, const V &v, data::function<V (const V &old_v, const V &new_v)> already_exists) const;
-
-            derived operator << (const entry<K, V> &e) const;
-
-            derived replace (const V &a, const V &b) const;
-            derived replace_part (const K &k, const V &v) const;
-            derived replace_part (const K &k, data::function<V (const V &)> f) const;
-
-            derived remove (const K &k) const;
-
-        };
-
-        template <typename derived, typename K, typename V, functional::map<K, V> M>
-        inline base_map<derived, K, V, M>::base_map (M &&rb) : M {rb} {}
-
-        template <typename derived, typename K, typename V, functional::map<K, V> M>
-        inline base_map<derived, K, V, M>::base_map (const M &rb) : M {rb} {}
-
-        template <typename derived, typename K, typename V, functional::map<K, V> M>
-        derived inline base_map<derived, K, V, M>::insert (const K &k, const V &v) const {
-            return derived {M::insert (k, v)};
-        }
-
-        template <typename derived, typename K, typename V, functional::map<K, V> M>
-        derived inline base_map<derived, K, V, M>::insert (const entry<K, V> &e) const {
-            return derived {M::insert (e)};
-        }
-
-        template <typename derived, typename K, typename V, functional::map<K, V> M>
-        derived inline base_map<derived, K, V, M>::insert (const K &k, const V &v,
-            data::function<V (const V &old_v, const V &new_v)> already_exists) const {
-            return derived {M::insert (k, v, already_exists)};
-        }
-
-        template <typename derived, typename K, typename V, functional::map<K, V> M>
-        derived inline base_map<derived, K, V, M>::operator << (const entry<K, V> &e) const {
-            return derived {M::insert (e)};
-        }
-
-        template <typename derived, typename K, typename V, functional::map<K, V> M>
-        derived inline base_map<derived, K, V, M>::replace (const V &a, const V &b) const {
-            return derived {M::replace (a, b)};
-        }
-
-        template <typename derived, typename K, typename V, functional::map<K, V> M>
-        derived inline base_map<derived, K, V, M>::replace_part (const K &k, const V &v) const {
-            return derived {M::replace_part (k, v)};
-        }
-
-        template <typename derived, typename K, typename V, functional::map<K, V> M>
-        derived inline base_map<derived, K, V, M>::replace_part (const K &k, data::function<V (const V &)> f) const {
-            return derived {M::replace_part (k, f)};
-        }
-
-        template <typename derived, typename K, typename V, functional::map<K, V> M>
-        derived inline base_map<derived, K, V, M>::remove (const K &k) const {
-            return derived {M::remove (k)};
-        }
     }
 
 }
