+++ src/data/include/data/tools.hpp
@@ -51,7 +51,6 @@ namespace data {
     // a functional map implemented as a red-black tree
     // wrapper of Milewski's implementation of Okasaki.
     template <typename K, typename V> using map = tool::rb_map<K, V>;
-    template <typename K, typename V, typename derived> using base_map = functional::base_map<derived, K, V, map<K, V>>;
     
     // set implemented as a map. 
     template <typename X> using set = tool::map_set<map<X, tool::unit>>;
